Extract helper methods to a separate class
